trading-bot/requirements.txt
fastapi
requests
python-dotenv
pydantic
schedule
websocket-client
cryptography
PyJWT
uvicorn